Change the return type of brlapi_getHandleSize() to size_t (from int) to match sizeof(). (st)
git-svn-id: svn://mielke.cc/main/brltty@2590 91a5dbb7-01b9-0310-9b5f-b28072856b6e

+++ Programs/api_client.c
@@ -184,7 +184,7 @@ struct brlapi_handle_t { /* Connection-specific information */
 };
 
 /* Function brlapi_getHandleSize */
-int brlapi_getHandleSize(void)
+size_t brlapi_getHandleSize(void)
 {
   return sizeof(brlapi_handle_t);
 }
